Keep the (inlined) function call in the loop
For some reason, Clang is too smart to ignore the fact that the function
call is a loop invariant, so it hoists the function *out of the loop*,
inlines it, and generates an empty (!) loop. This of course leads to
bogus results.

As a short-term band aid, fiddle around with the function call argument
to convince Clang to keep the benchmark subject in the loop. This still
involves eyeballing the generated assemblyl, so yuck. I don't
know what a longer term fix looks like, preferably it should not involve
an extra memory load interleaved with the inlined code.